WIP: partially migrated the dataset_tree_info to services
